backend/src/requirements.txt
Flask==2.2.2
Flask-Caching==2.0.2
Pillow==9.4.0
scipy==1.10.0
numpy==1.24.1
Werkzeug==2.2.2
flask-cors==3.0.10